docker-compose orchestration for Elasticsearch cluster
Separate data volumes are not included in this build; these don't work when we
scale them up via `docker-compose` because stopped data volumes will be removed
when we scale up. This also removes the necessity of worrying about collocation.

For discovery we're going to take advantage of a fairly typical production
topology for Elasticsearch which is master-only nodes. Our data nodes will be
linked to master nodes and be able to use that container's IP to bootstrap
unicast zen discovery.